+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 Latest
 ------
+* Added APILoader.load_location() for adding or updating database records.
 
 0.6.4 (2025-01-27)
 ------------------